[Issue-4094] refactor code
