+++ packages/extension-koni-ui/src/components/AccountProxy/index.ts
@@ -9,6 +9,7 @@ export { default as AccountProxyItem } from './AccountProxyItem';
 export { default as AccountProxyAvatar } from './AccountProxyAvatar';
 export { default as AccountProxyAvatarGroup } from './AccountProxyAvatarGroup';
 export { default as AccountChainAddressItem } from './AccountChainAddressItem';
+export { default as AccountTokenAddressItem } from './AccountTokenAddressItem';
 export { default as AddressSelectorItem } from './AddressSelectorItem';
 export { default as AccountProxyTypeTag } from './AccountProxyTypeTag';
 export { default as AccountChainAddressWithStatusItem } from './AccountChainAddressWithStatusItem';
