+++ packages/extension-koni-ui/src/types/account.ts
@@ -3,7 +3,7 @@
 
 import type { KeypairType } from '@subwallet/keyring/types';
 
-import { AccountActions, AccountProxyType } from '@subwallet/extension-base/types';
+import { AccountActions, AccountJson, AccountProxyType } from '@subwallet/extension-base/types';
 
 export interface WordItem {
   index: number;
@@ -34,6 +34,12 @@ export type AccountChainAddress = {
   address: string;
   accountType: KeypairType;
   logoKey?: string
+}
+
+export type AddressGroupItemInfo = {
+  accountInfo: AccountJson;
+  tokenSlug: string;
+  chainSlug: string;
 }
 
 export interface BitcoinAccountInfo {
